+++ src/main/java/com/yourdomain/guardianac/checks/combat/nohitdelay/NoHitDelayC.java
@@ -34,7 +34,7 @@ public class NoHitDelayC extends Check {
         PlayerData data = getPlugin().getPlayerDataManager().getPlayerData(player);
         if (data == null) return;
 
-        List<Long> intervals = data.getAttackIntervals();
+        List<Double> intervals = data.getAttackIntervals();
         if (intervals.size() < MIN_SAMPLES) return;
 
         double mean = MathUtil.mean(intervals);
